Fix code duplication issue
