+++ simple-loadflow/src/main/java/com/powsybl/loadflow/simple/network/LfDanglingLineBus.java
@@ -7,43 +7,24 @@
 package com.powsybl.loadflow.simple.network;
 
 import com.powsybl.iidm.network.DanglingLine;
-import com.powsybl.iidm.network.ShuntCompensator;
 
-import java.util.Collections;
-import java.util.List;
 import java.util.Objects;
 
 /**
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
  */
-public class LfDanglingLineBus extends AbstractLfBus {
+public class LfDanglingLineBus extends AbstractFictitiousLfBus {
 
     private final DanglingLine danglingLine;
 
-    private final int num;
-
-    private double v = Double.NaN;
-
-    private double angle = Double.NaN;
-
     public LfDanglingLineBus(DanglingLine danglingLine, int num) {
+        super(num);
         this.danglingLine = Objects.requireNonNull(danglingLine);
-        this.num = num;
     }
 
     @Override
     public String getId() {
         return danglingLine.getId() + "_BUS";
-    }
-
-    @Override
-    public int getNum() {
-        return num;
-    }
-
-    @Override
-    public boolean hasVoltageControl() {
-        return false;
     }
 
     @Override
@@ -57,48 +38,8 @@ public class LfDanglingLineBus extends AbstractLfBus {
     }
 
     @Override
-    public double getGenerationTargetP() {
-        return 0;
-    }
-
-    @Override
-    public double getGenerationTargetQ() {
-        return 0;
-    }
-
-    @Override
-    public double getTargetV() {
-        return Double.NaN;
-    }
-
-    @Override
-    public double getV() {
-        return v;
-    }
-
-    @Override
-    public void setV(double v) {
-        this.v = v;
-    }
-
-    @Override
-    public double getAngle() {
-        return angle;
-    }
-
-    @Override
-    public void setAngle(double angle) {
-        this.angle = angle;
-    }
-
-    @Override
     public double getNominalV() {
         return danglingLine.getTerminal().getVoltageLevel().getNominalV();
-    }
-
-    @Override
-    public List<ShuntCompensator> getShuntCompensators() {
-        return Collections.emptyList();
     }
 
     @Override
